+++ src/App/backend/src/Altinn.App.Api/Controllers/ProcessController.cs
@@ -37,7 +37,6 @@ public class ProcessController : ControllerBase
     private readonly IInstanceClient _instanceClient;
     private readonly IProcessClient _processClient;
     private readonly IProcessEngine _processEngine;
-    private readonly IProcessReader _processReader;
     private readonly IProcessEngineAuthorizer _processEngineAuthorizer;
     private readonly IValidationService _validationService;
     private readonly InstanceDataUnitOfWorkInitializer _instanceDataUnitOfWorkInitializer;
@@ -54,7 +53,6 @@ public class ProcessController : ControllerBase
         IInstanceClient instanceClient,
         IProcessClient processClient,
         IValidationService validationService,
-        IProcessReader processReader,
         IServiceProvider serviceProvider,
         IProcessEngineAuthorizer processEngineAuthorizer,
         ProcessStateEnricher processStateEnricher
@@ -63,7 +61,6 @@ public class ProcessController : ControllerBase
         _logger = logger;
         _instanceClient = instanceClient;
         _processClient = processClient;
-        _processReader = processReader;
         _processEngine = serviceProvider.GetRequiredService<IProcessEngine>();
         _processEngineAuthorizer = processEngineAuthorizer;
         _validationService = validationService;
@@ -233,76 +230,6 @@ public class ProcessController : ControllerBase
             return ExceptionResponse(
                 startException,
                 $"Unable to start the process for instance {instance?.Id} of {instance?.AppId}"
-            );
-        }
-    }
-
-    /// <summary>
-    /// Gets a list of the next process elements that can be reached from the current process element.
-    /// If process is not started it returns the possible start events.
-    /// </summary>
-    /// <param name="org">unique identifier of the organization responsible for the app</param>
-    /// <param name="app">application identifier which is unique within an organization</param>
-    /// <param name="instanceOwnerPartyId">unique id of the party that is the owner of the instance</param>
-    /// <param name="instanceGuid">unique id to identify the instance</param>
-    /// <returns>list of next process element identifiers (tasks or events)</returns>
-    [Authorize(Policy = AuthzConstants.POLICY_INSTANCE_READ)]
-    [HttpGet("next")]
-    [ProducesResponseType(StatusCodes.Status200OK)]
-    [ProducesResponseType(StatusCodes.Status409Conflict)]
-    [Obsolete(
-        "From v8 of nuget package navigation is done by sending performed action to the next api. Available actions are returned in the GET /process endpoint"
-    )]
-    public async Task<ActionResult<List<string>>> GetNextElements(
-        [FromRoute] string org,
-        [FromRoute] string app,
-        [FromRoute] int instanceOwnerPartyId,
-        [FromRoute] Guid instanceGuid
-    )
-    {
-        Instance? instance = null;
-        string? currentTaskId = null;
-
-        try
-        {
-            instance = await _instanceClient.GetInstance(
-                app,
-                org,
-                instanceOwnerPartyId,
-                instanceGuid,
-                authenticationMethod: null,
-                CancellationToken.None
-            );
-
-            if (instance.Process == null)
-            {
-                return Ok(_processReader.GetStartEventIds());
-            }
-
-            currentTaskId = instance.Process.CurrentTask?.ElementId;
-
-            if (currentTaskId == null)
-            {
-                return Conflict($"Instance does not have valid info about currentTask");
-            }
-
-            return Ok(new List<string>());
-        }
-        catch (PlatformHttpException e)
-        {
-            return HandlePlatformHttpException(
-                e,
-                $"Unable to find next process element for instance {instance?.Id} and current task {currentTaskId}. Exception was {e.Message}. Is the process file OK?"
-            );
-        }
-        catch (Exception processException)
-        {
-            _logger.LogError(
-                $"Unable to find next process element for instance {instance?.Id} and current task {currentTaskId}. {processException}"
-            );
-            return ExceptionResponse(
-                processException,
-                $"Unable to find next process element for instance {instance?.Id} and current task {currentTaskId}. Exception was {processException.Message}. Is the process file OK?"
             );
         }
     }